requirements.txt
openai==0.27.6
tiktoken==0.3.3
python-dotenv==1.0.0
